requirements.txt
# === Voice Phishing Detection System Requirements ===

# FastAPI 및 웹 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# LangChain 및 AI 모델
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-core==0.1.10
langchain-teddynote

# LangGraph
langgraph==0.0.20

# OpenAI
openai==1.3.7

# Cohere (Reranker)
cohere==4.37

# 벡터 DB 및 검색
faiss-cpu==1.7.4
pydantic==2.5.0

# 환경 변수 관리
python-dotenv==1.0.0

# HTTP 요청 (STT API)
requests==2.31.0

# 데이터 처리
numpy==1.24.3
pandas==2.0.3

# 기타 유틸리티
python-multipart==0.0.6
aiofiles==23.2.1

